pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "coherent-client"
version = "0.0.3"
authors = [
  { name="coherentvolition", email="coherentvolition@protonmail.com" },
]
description = "An API client for the coherent network that implements authentication and general requests"
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = ["requests", "pyjwt[crypto]", "python-dotenv"]

[project.urls]
Homepage = "https://github.com/Coherent-Volition-Inc/coherent-client"
Issues = "https://github.com/Coherent-Volition-Inc/coherent-client"